Trailing elements of a list scraper instruction act as filters
The filter can be a function, or can it be another instruction. For instance, if
you want to only match tables with rows:

["tr" "td"]

Or only match links containing certain text:

["a" (text, re.compile(r"certain text"))]

(Regular expression instructions work as expected.)

Note that instructions after the first one inside [] will refine the initial
selection but not expand it or drill into it. That's the purpose of ()
instructions.